day 7 part 1 with functional backtracking

+++ projects/day7/lib.h
@@ -1,4 +1,4 @@
 #pragma once
 #include "./parser.h"
 
-int part1(char *input_path);
+unsigned long part1(char *input_path);
